django-blog-master/django-blog-master/django-blog-master/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-bootstrap3==12.0.3
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0.1
django-js-asset==1.2.2
django-mdeditor==0.1.9
django-ranged-response==0.2.0
django-registration==3.0.1
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.0
jdcal==1.4.1
lxml==4.4.2
Markdown==3.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
Pygments==2.5.2
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
six==1.14.0
soupsieve==1.9.5
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
